Seperate types from libs
