+++ archetypes/ArbitrumBridge/index.tsx
@@ -5,8 +5,8 @@ import { useWeb3 } from '@context/Web3Context/Web3Context';
 import { MultiBridge } from './MultiBridge';
 import { switchNetworks } from '@libs/utils/rpcMethods';
 import { Network } from '@context/Web3Context/Web3Context.Config';
-import { BridgeableAsset } from '@libs/types/General';
-import { bridgeableTickers } from '@libs/utils/bridge';
+import { BridgeableAsset } from '@/types/bridge';
+import { bridgeableTickers } from '@constants/bridge';
 
 // ArbitrumBridge
 export const ArbitrumBridge: React.FC = (() => {
