+++ types/filters.ts
@@ -1,0 +1,9 @@
+export enum MarketFilterEnum {
+    All = 'All',
+    ETH = 'Ethereum',
+    BTC = 'Bitcoin',
+    TOKE = 'Tokemak',
+    LINK = 'Chainlink',
+    EUR = 'Euro',
+    AAVE = 'Aave',
+}
