+++ types/staking/balancerV2Vault/Vault.d.ts
