+++ types/staking/typechain/ERC20Detailed.d.ts
