+++ types/staking/typechain/IStakingRewards.d.ts
