+++ types/staking/typechain/MockToken.d.ts
