+++ types/staking/typechain/Owned.d.ts
