+++ types/staking/typechain/StakingRewards.d.ts
