+++ types/staking/typechain/factories/IStakingRewards__factory.ts
