+++ types/staking/typechain/factories/Owned__factory.ts
