+++ types/staking/typechain/factories/StakingRewards__factory.ts
